Fixed Nightly CIS errors
#rb trivial
#jira none

[CL 4688073 by Mattias Hornlund in Dev-Networking branch]

+++ Engine/Source/Runtime/Engine/Public/Net/NetPacketNotify.h
@@ -124,7 +124,7 @@ public:
 
 	SequenceNumberT UpdateInAckSeqAck(SequenceNumberT::DifferenceT AckCount, SequenceNumberT AckedSeq)
 	{
-		check(AckCount <= AckRecord.Count());
+		check((SIZE_T)AckCount <= AckRecord.Count());
 
 		if (AckCount > 1)
 		{
